Сделал рефакторинг module5-task1
