+++ js/type.js
@@ -1,0 +1,20 @@
+/**
+ * @typedef CommentState
+ * @prop {number} id
+ * @prop {string} avatar
+ * @prop {string} message
+ * @prop {string} name
+ */
+
+/**
+ * @typedef ImageState
+ * @prop {number} id
+ * @prop {string} url
+ * @prop {string} description
+ * @prop {number} likes
+ * @prop {CommentState[]} comments
+ */
+
+/**
+ * @typedef {[min: number, max: number]} NumberRange
+ */
